+++ Fenetres/Saut.cs
@@ -2,7 +2,7 @@
 using System.Windows.Forms;
 using Gfe.Langues;
 
-namespace Gfe
+namespace Gfe.Fenetres
 {
     public partial class Saut : Form
     {