+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ﻿
 ## Changelog du Gestionnaire de Fond d'Écran
 
-### Version 0.9.2 (à venir) :
+### Version 0.9.2 (4 Septembre 2017) :
  - Réorganisation interne des fichiers.
  - Passage au framework .NET 4.7
  - Ajout d'une option pour obtenir la liste des fichiers et pour sélectionner un fichier précis dans la liste.